pyproject.toml
[tool.poetry]
name = "nymbleup-assignment"
version = "0.1.0"
description = ""
authors = ["Satyarth <satyarthdev@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.11"
psycopg2-binary = "^2.9.3"
django-cors-headers = "^3.11.0"
drf-yasg = "^1.20.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
